second commit

+++ src/interfaces/ITreasury.sol
@@ -4,13 +4,10 @@ pragma solidity 0.7.5;
 
 interface ITreasury {
     function deposit(
-        address _principleTokenAddress, 
-        uint _amountPrincipleToken, 
-        uint _amountPayoutToken
+        address _principleTokenAddress,
+        uint256 _amountPrincipleToken,
+        uint256 _amountPayoutToken
     ) external;
 
-    function valueOfToken(
-        address _principleTokenAddress, 
-        uint _amount
-    ) external view returns ( uint value_ );
+    function valueOfToken(address _principleTokenAddress, uint256 _amount) external view returns (uint256 value_);
 }
